Use slack-webhook library
We were previously building the json payload ourselves. Now, we off-load
that work onto a library which is made for this purpose.

This is just an initial shim that works. We probably need to ditch the
SlackPayload class entirely favoring a SlackMessage Factory that does
the work currently in the SlackPayload constructor.